requirements.txt
affinegap==1.11
alabaster==0.7.12
asyncio==3.4.3
attrs==21.2.0
Babel==2.9.1
base58==2.1.0
bleach==4.1.0
canonicaljson==1.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
colorclass==2.2.0
cryptography==35.0.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
doxypypy==0.8.8.6
fett==0.3.2
flake8==4.0.1
frozendict==2.0.6
idna==3.3
imagesize==1.2.0
importlib-metadata==4.8.1
iniconfig==1.1.1
javaobj-py3==0.4.3
jeepney==0.7.1
Jinja2==3.0.2
keyring==23.2.1
lib==3.0.0
lru-dict==1.1.7
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
more-itertools==8.10.0
networkx==2.6.3
numpy==1.21.2
packaging==21.0
pip-review==1.1.0
pip-upgrade==0.0.6
pip-upgrader==1.4.15
pipdeptree==2.2.0
pkginfo==1.7.1
pluggy==1.0.0
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pycryptodomex==3.11.0
pyflakes==2.4.0
Pygments==2.10.0
pyjks==20.0.0
pymongo==3.12.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.5
python-axolotl-curve25519==0.4.1.post2
python-dotenv==0.19.1
python-jsonrpc-server==0.4.0
pytz==2021.3
PyYAML==6.0
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
simplejson==3.17.5
six==1.16.0
snooty-lextudio==1.11.4.dev0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
terminaltables==3.1.0
toml==0.10.2
tqdm==4.62.3
twine==3.4.2
twofish==0.3.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.6.0

setuptools~=58.2.0